+++ R/tbl_propdiff.R
@@ -10,6 +10,10 @@
 #' non-missing levels
 #' @param formula By default, `"{y} ~ {x}"`. To include covariates for an adjusted
 #' risk difference, add covariate names to the formula, e.g. `"{y} ~ {x} + age"`
+#' @param label List of formulas specifying variables labels, If a variable's label is
+#' not specified here, the label attribute (`attr(data$high_grade, "label")`) is used.
+#' If attribute label is `NULL`, the variable name will be used.
+#' @param statistic Statistics to display for each group. Default `"{n} ({p}%)"`
 #' @param method The method for calculating p-values and confidence intervals around the
 #' difference in rates. The options are `"chisq"`, `"exact"`, `"boot_centile"`,
 #' and `"boot_sd"`. See below for details. Default method is `"chisq"`.
@@ -39,13 +43,14 @@
 #' in all bootstrap samples (the default for this method is 2000 resamples)
 #' and generates the confidence intervals from the distribution of these
 #' differences. For the default, a 95% confidence interval, the 2.5 and 97.5
-#' centiles are used.
+#' centiles are used. The p-value presented is from a logistic regression model.
 #'
 #' * The `boot_sd` option calculates the adjusted difference between groups
 #' in all bootstrap samples (the default for this method is 200 resamples).
 #' The mean and standard deviation of the adjusted difference across all
 #' resamples are calculated. The standard deviation is then used as the
-#' standard error to calculate the confidence interval based on the true adjusted difference.
+#' standard error to calculate the confidence interval based on the true
+#' adjusted difference. The p-value presented is from a logistic regression model.
 #'
 #' @examples
 #' tbl_propdiff(
@@ -62,12 +67,20 @@
 #'   method = "boot_sd",
 #'   bootstrapn = 25
 #' )
-tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
+tbl_propdiff <- function(data, y, x,
+                         formula = "{y} ~ {x}",
+                         label = NULL,
+                         statistic = "{n} ({p}%)",
                          method = c("chisq", "exact", "boot_sd", "boot_centile"),
                          conf.level = 0.95,
                          bootstrapn = ifelse(method == "boot_centile", 2000, 200),
                          estimate_fun = style_sigfig,
                          pvalue_fun = style_pvalue) {
+
+  # converting inputs to named list --------------------------------------------
+  y <- var_input_to_string(data = data, select_input = {{ y }}, arg_name = "y", select_single = FALSE)
+  x <- var_input_to_string(data = data, select_input = {{ x }}, arg_name = "x", select_single = TRUE)
+  label <- tidyselect_to_list(.data = data, x = {{ label }}, arg_name = "label", select_single = FALSE)
 
   ### CHECKS------------------
   # Matching arguments for method
@@ -81,10 +94,18 @@ tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
   if (length(setdiff(c(x, y, covariates), names(data))) != 0) {
     stop(glue(
       "These variables do not exist in the dataset: ",
-      glue_collapse(setdiff(c(x, y), names(data)), sep = ", ")
+      glue_collapse(setdiff(c(x, y, covariates), names(data)), sep = ", ")
     ),
     call. = FALSE
     )
+  }
+
+  # If specifying covariates but not specifying a multivariable method
+  if (length(covariates) != 0 & !(method %in% c("boot_centile", "boot_sd"))) {
+    stop(glue(
+      "When specifying a formula that includes covariates, you must specify one of the two multivariable methods ('boot_centile' or 'boot_sd')."
+    ),
+    call. = FALSE)
   }
 
   # If selecting a multivariable method but no covariates
@@ -93,10 +114,6 @@ tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
       "Method '{method}' was selected but no covariates were provided. ",
       "The {conf.level*100}% confidence interval around the unadjusted difference in rates will be bootstrapped."))
     }
-
-  # converting inputs to strings/lists
-  y <- dplyr::select(data[0, , drop = FALSE], {{ y }}) %>% names()
-  x <- dplyr::select(data[0, , drop = FALSE], {{ x }}) %>% names()
 
   # checking the x variable has two levels
   if (data[[x]] %>% stats::na.omit() %>% unique() %>% length() != 2) {
@@ -143,7 +160,6 @@ tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
   data <-
     data %>%
     dplyr::mutate_if(is.logical, as.numeric) %>%
-    mutate_at(vars(x), ~ forcats::fct_rev(factor(.))) %>%
     mutate_at(vars(y), ~ factor(.))
 
   ### UNADJUSTED RATES---------------------------------
@@ -155,10 +171,7 @@ tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
       outcome_label =
         map_chr(
           y,
-          ~ ifelse(!is.null(attr(data[[..1]], "label")),
-            attr(data[[..1]], "label"),
-            y
-          )
+          ~ label[[..1]] %||% attr(data[[..1]], "label") %||% ..1
         ),
       # Save out table of unadjusted rates
       tbl_rates =
@@ -167,10 +180,12 @@ tbl_propdiff <- function(data, y, x, formula = "{y} ~ {x}",
           function(x, y, z) {
             data %>%
               select(all_of(c(x, y))) %>%
+              mutate_at(vars(any_of(x)), ~factor(.) %>% forcats::fct_rev()) %>%
               gtsummary::tbl_summary(
                 by = .data[[x]], missing = "no",
-                label = list(x = glue("{z}")),
-                type = list(all_categorical() ~ "dichotomous")
+                label = list(z) %>% rlang::set_names(y),
+                type = list(everything() ~ "dichotomous"),
+                statistic = list(everything() ~ glue("{statistic}"))
               ) %>%
               gtsummary::add_n() %>%
               gtsummary::modify_header(stat_by = "**{level}**")